code-templates/financial-services-kb-template/app/document_processing/document_processor.py
import os
from typing import Dict, List, Optional, Tuple
import PyPDF2
from docx import Document as DocxDocument
import json
from pathlib import Path

class DocumentProcessor:
    """
    Processes documents for the knowledge base.
    
    This class handles document ingestion, text extraction, and embedding generation.
    Developers will implement this in Milestone 2.
    """
    
    def __init__(self, raw_dir: str, processed_dir: str, embeddings_dir: str):
        """
        Initialize the document processor.
        
        Args:
            raw_dir: Directory containing raw documents
            processed_dir: Directory to store processed documents
            embeddings_dir: Directory to store document embeddings
        """
        self.raw_dir = raw_dir
        self.processed_dir = processed_dir
        self.embeddings_dir = embeddings_dir
        
        # Create directories if they don't exist
        for directory in [self.raw_dir, self.processed_dir, self.embeddings_dir]:
            os.makedirs(directory, exist_ok=True)
    
    def process_document(self, file_path: str) -> Dict:
        """
        Process a document and extract its content and metadata.
        
        This method will be implemented by developers in Milestone 2.
        
        Args:
            file_path: Path to the document file
            
        Returns:
            Dictionary containing document content and metadata
        """
        # Placeholder implementation
        file_name = os.path.basename(file_path)
        return {
            "id": file_name,
            "title": file_name,
            "content": "Sample content",
            "metadata": {
                "access_level": "public",
                "department": "finance",
                "type": "policy"
            }
        }
    
    def extract_text(self, file_path: str) -> str:
        """
        Extract text from a document file.
        
        This method will be implemented by developers in Milestone 2.
        
        Args:
            file_path: Path to the document file
            
        Returns:
            Extracted text content
        """
        # Placeholder implementation
        return "Sample document text"
    
    def generate_embeddings(self, text: str) -> List[float]:
        """
        Generate embeddings for document text.
        
        This method will be implemented by developers in Milestone 2.
        
        Args:
            text: Document text
            
        Returns:
            Vector embedding of the text
        """
        # Placeholder implementation - returns a small random vector
        import random
        return [random.random() for _ in range(10)]
    
    def process_all_documents(self) -> List[Dict]:
        """
        Process all documents in the raw directory.
        
        This method will be implemented by developers in Milestone 2.
        
        Returns:
            List of processed document dictionaries
        """
        # Placeholder implementation
        return [
            {
                "id": "doc1.pdf",
                "title": "Financial Policy 2023",
                "content": "Sample content for financial policy",
                "metadata": {
                    "access_level": "restricted",
                    "department": "finance",
                    "type": "policy"
                }
            },
            {
                "id": "doc2.pdf",
                "title": "Compliance Guidelines",
                "content": "Sample content for compliance guidelines",
                "metadata": {
                    "access_level": "public",
                    "department": "compliance",
                    "type": "guidelines"
                }
            }
        ]

# Create a singleton instance with default directories
document_processor = DocumentProcessor(
    raw_dir="data/raw",
    processed_dir="data/processed",
    embeddings_dir="data/embeddings"
)
